Add auth service which interacts with engine
Now we have a clear interface for the biometrics engine, I added a
dedicated `auth` service which is acts as a central point for all
auth-related queries. For our application, it stores the current
authentication state in a cookie which Twilio will forward to all
subsequent steps in the call.