Separate out engine and voiceIt provider
I wanted to add a clearer separation between the app and the engine. Now
they are clearly divided, with an explicit interface between them. The
interface has four methods: `getEnrolmentRequest`,
`handleEnrolmentResponse`, `getVerficiationRequest`,
`handleVerificationResponse`. All future biometrics providers should be
able to fill this interface (possibly with minor tweaks to the interface
along the way). For now, we just have an implementation for a VoiceIT
provider.