Add params to safeHandle to enable auth interception
The `safeHandle` function was already used to simplify writing Twilio
handlers by providing common serialisation/deserialisation logic. I've
added parameters to the function which also allows it to add common
authentication logic. It's now possible to add a `requireVerification`
parameter which will cause `safeHandle` to intercept requests to the
handler if the user has not yet been authenticated. If the number is
recognised as previously enroled, `safeHandle` will redirect to a
verification flow, or to an enrolment flow if not. After the
verification flow is completed, `safeHandle` will redirect the user back
to the handler they were originally expecting to hit.