Use new auth on menu handler and update strings
Now that we have added the `requireVerification` parameter, it's time to
use it. Here I simply add it to the menu handler to prevent access to
the account menu before the user's voice has been verified. I also
updated some strings to make the flow slightly more natural, and added
missing languages.