Separated Biometrics Engine  (#27)
* Separate out engine and voiceIt provider

I wanted to add a clearer separation between the app and the engine. Now
they are clearly divided, with an explicit interface between them. The
interface has four methods: `getEnrolmentRequest`,
`handleEnrolmentResponse`, `getVerficiationRequest`,
`handleVerificationResponse`. All future biometrics providers should be
able to fill this interface (possibly with minor tweaks to the interface
along the way). For now, we just have an implementation for a VoiceIT
provider.

* Add auth service which interacts with engine

Now we have a clear interface for the biometrics engine, I added a
dedicated `auth` service which is acts as a central point for all
auth-related queries. For our application, it stores the current
authentication state in a cookie which Twilio will forward to all
subsequent steps in the call.

* Remove enrolment and verification routes

Within our call center application, having dedicated routes for
enrolment and verification was actually a little limiting. In
particular, it makes it difficult to have multiple entry-points
(necessary for alert callbacks) without duplicating the verification
routes. Here I removed the dedicated routes, to replace them later with
a different solution.

* Add params to safeHandle to enable auth interception

The `safeHandle` function was already used to simplify writing Twilio
handlers by providing common serialisation/deserialisation logic. I've
added parameters to the function which also allows it to add common
authentication logic. It's now possible to add a `requireVerification`
parameter which will cause `safeHandle` to intercept requests to the
handler if the user has not yet been authenticated. If the number is
recognised as previously enroled, `safeHandle` will redirect to a
verification flow, or to an enrolment flow if not. After the
verification flow is completed, `safeHandle` will redirect the user back
to the handler they were originally expecting to hit.

* Use new auth on menu handler and update strings

Now that we have added the `requireVerification` parameter, it's time to
use it. Here I simply add it to the menu handler to prevent access to
the account menu before the user's voice has been verified. I also
updated some strings to make the flow slightly more natural, and added
missing languages.

* Update tests