Simplify implementation by allowing handlers to set cookies
During the process of adding new enrolments, we don't need to mess about with
authentication, because we already have a valid set of enrolments in place.
This means that our "add passphrase" handlers can be independent and not worry
about accidentally breaking something inside the authentication flow (and it
makes things simpler not to worry about auth!).

There is some duplication of logic (especially the `requestEnrolment` and
`handleEnrolment` functions, but it's tolerable. We wouldn't want to reuse
those functions directly because they make other changes as well as creating
enrolments (e.g. creating a new voiceit user). In an ideal world, we would
extract the common logic into an independent function, and call that logic in
two different places (`handleEnrolment` and the "add passphrase" handler).

The remaining missing piece is that we need some way of storing the original
request between the moment when we ask the user for a voice recording, and the
moment that we receive the new recording. This can be tricky in a serverless
environment, because those two moments might occur on totally different
machines! There are two solutions available:

1. Add the data that needs preserving into DynamoDB
2. Add the data into an HTTP cookie on the first request, which Twilio will send back to us for the second.

Currently the authentication flow uses the 2nd approach, to allow DynamoDB to
exclusively represent the "business logic" of our call center application, and
keep it isolated from the implementation details of our biometric
authentication. Unfortunately function handlers didn't previously have a way of
setting cookies, all they can do is return a voiceIt `response` object which
gets turned into an HTTP response.  I've added this ability, so that any
function handlers can now return a response body along with a record containing
cookies.